chore: separate pallet logic to the modules
