+++ pallets/storage-provider/src/hooks/mod.rs
@@ -1,0 +1,5 @@
+mod check_deadlines;
+mod check_precommited_sectors;
+
+pub use check_deadlines::check_deadlines;
+pub use check_precommited_sectors::check_precommited_sectors;
