+++ pallets/storage-provider/src/lib.rs
@@ -15,9 +15,11 @@
 mod tests;
 
 mod deadline;
+mod dispatchables;
 pub mod error;
 pub mod expiration_queue;
 pub mod fault;
+mod hooks;
 mod partition;
 pub mod proofs;
 pub mod sector;
@@ -29,59 +31,38 @@ pub use pallet::*;
 #[frame_support::pallet(dev_mode)]
 pub mod pallet {
     pub(crate) const DECLARATIONS_MAX: u32 = 3000;
-    const LOG_TARGET: &'static str = "runtime::storage_provider";
+    pub(crate) const LOG_TARGET: &'static str = "runtime::storage_provider";
 
     extern crate alloc;
 
-    use alloc::{collections::BTreeMap, vec, vec::Vec};
+    use alloc::vec;
     use core::fmt::Debug;
 
-    use cid::Cid;
     use codec::{Decode, Encode};
-    use frame_support::{
-        dispatch::DispatchResult,
-        ensure, fail,
-        pallet_prelude::*,
-        sp_runtime::traits::{CheckedAdd, CheckedSub, One},
-        traits::Randomness,
-    };
-    use frame_system::{
-        ensure_signed,
-        pallet_prelude::{BlockNumberFor, *},
-    };
+    use frame_support::{dispatch::DispatchResult, pallet_prelude::*, traits::Randomness};
+    use frame_system::pallet_prelude::{BlockNumberFor, *};
     use primitives::{
-        commitment::{CommD, CommR, Commitment},
         configs::{BalanceOf, CurrencyProvider, StorageProviderProvider},
         pallets::{
             DeadlineInfo as ExternalDeadlineInfo, Market, ProofVerification,
             StorageProviderValidation,
         },
-        proofs::{
-            assign_proving_period_offset, derive_prover_id, PublicReplicaInfo, RegisteredPoStProof,
-        },
-        randomness::{draw_randomness, AuthorVrfHistory, DomainSeparationTag},
+        proofs::RegisteredPoStProof,
+        randomness::AuthorVrfHistory,
         sector::{ProveCommitSector, SectorNumber, SectorPreCommitInfo},
-        PartitionNumber, MAX_DEALS_PER_SECTOR, MAX_PARTITIONS_PER_DEADLINE,
-        MAX_POREP_PROOFS_PER_BLOCK, MAX_SEAL_PROOF_BYTES, MAX_SECTORS, MAX_SECTORS_PER_CALL,
+        PartitionNumber, MAX_PARTITIONS_PER_DEADLINE, MAX_SECTORS, MAX_SECTORS_PER_CALL,
     };
     use scale_info::TypeInfo;
-    use sp_arithmetic::traits::Zero;
 
     use crate::{
-        deadline::{deadline_is_mutable, DeadlineInfo},
+        deadline::DeadlineInfo,
         fault::{
             DeclareFaultsParams, DeclareFaultsRecoveredParams, FaultDeclaration,
             RecoveryDeclaration,
         },
         proofs::SubmitWindowedPoStParams,
-        sector::{
-            ProveCommitResult, SectorOnChainInfo, SectorPreCommitOnChainInfo,
-            TerminateSectorsParams, TerminationDeclaration,
-        },
-        sector_map::DeadlineSectorMap,
-        storage_provider::{
-            calculate_first_proving_period_start, StorageProviderInfo, StorageProviderState,
-        },
+        sector::{ProveCommitResult, TerminateSectorsParams, TerminationDeclaration},
+        storage_provider::{StorageProviderInfo, StorageProviderState},
     };
 
     #[pallet::pallet]
@@ -248,7 +229,7 @@ pub mod pallet {
     >;
 
     #[pallet::event]
-    #[pallet::generate_deposit(fn deposit_event)]
+    #[pallet::generate_deposit(pub(super) fn deposit_event)]
     pub enum Event<T: Config> {
         /// Emitted when a new storage provider is registered.
         StorageProviderRegistered {
@@ -386,43 +367,11 @@ pub mod pallet {
             peer_id: T::PeerId,
             window_post_proof_type: RegisteredPoStProof,
         ) -> DispatchResult {
-            let owner = ensure_signed(origin)?;
-            // Ensure that the storage provider does not exist yet
-            ensure!(
-                !StorageProviders::<T>::contains_key(&owner),
-                Error::<T>::StorageProviderExists
-            );
-            let current_block = <frame_system::Pallet<T>>::block_number();
-            let proving_period = T::WPoStProvingPeriod::get();
-
-            let offset = assign_proving_period_offset::<T::AccountId, BlockNumberFor<T>>(
-                &owner,
-                current_block,
-                T::WPoStProvingPeriod::get(),
-            );
-
-            let local_proving_start = calculate_first_proving_period_start::<BlockNumberFor<T>>(
-                current_block,
-                offset,
-                proving_period,
-            );
-            let info = StorageProviderInfo::new(peer_id, window_post_proof_type);
-            let state = StorageProviderState::<T::PeerId, BalanceOf<T>, BlockNumberFor<T>>::new(
-                info.clone(),
-                local_proving_start,
-                // Always zero since we're calculating the absolute first start
-                // thus the deadline will always be zero
-                0,
-                T::WPoStPeriodDeadlines::get(),
-            );
-            StorageProviders::<T>::insert(&owner, state);
-            // Emit event
-            Self::deposit_event(Event::StorageProviderRegistered {
-                owner,
-                info,
-                proving_period_start: local_proving_start,
-            });
-            Ok(())
+            crate::dispatchables::register_storage_provider::<T>(
+                origin,
+                peer_id,
+                window_post_proof_type,
+            )
         }
 
         /// The Storage Provider uses this extrinsic to pledge and seal X sectors at once.
@@ -442,125 +391,7 @@ pub mod pallet {
                 ConstU32<MAX_SECTORS_PER_CALL>,
             >,
         ) -> DispatchResult {
-            let owner = ensure_signed(origin)?;
-            let sp = StorageProviders::<T>::try_get(&owner)
-                .map_err(|_| Error::<T>::StorageProviderNotFound)?;
-            let current_block = <frame_system::Pallet<T>>::block_number();
-            let sector_amount = sectors.len();
-
-            // Pre-committed sectors for emitting the event.
-            let mut pre_committed_sectors = BoundedVec::new();
-            // All sectors in the batch to avoid mutating the SP multiple times.
-            let mut on_chain_sectors: BoundedVec<
-                SectorPreCommitOnChainInfo<BalanceOf<T>, BlockNumberFor<T>>,
-                ConstU32<MAX_SECTORS_PER_CALL>,
-            > = BoundedVec::new();
-            // Total deposit amount to avoid mutating the SP multiple times and reserve only once.
-            let mut total_deposit = BalanceOf::<T>::zero();
-            // sector deals for all pre commits
-            let mut all_sector_deals = BoundedVec::new();
-            // unsealed_cids for all sectors
-            let mut unsealed_cids = BoundedVec::new();
-            // deal amounts for each sector
-            let mut deal_amounts = BoundedVec::new();
-
-            for sector in sectors {
-                // Basic pre-commit validation.
-                Self::validate_sector_for_pre_commit(&sp, &sector)?;
-
-                // Check that the expiration set by the SP makes sense.
-                Self::validate_expiration(
-                    current_block,
-                    current_block + T::MaxProveCommitDuration::get(),
-                    sector.expiration,
-                )?;
-
-                // Validate the data commitment
-                let commd = Commitment::<CommD>::from_cid_bytes(&sector.unsealed_cid[..])
-                    .map_err(|err| {
-                        log::error!(target: LOG_TARGET, err:?; "pre_commit_sectors: invalid unsealed_cid");
-                        Error::<T>::InvalidCid
-                    })?;
-
-                // Validate the replica commitment
-                let _ = Commitment::<CommR>::from_cid_bytes(&sector.sealed_cid[..])
-                    .map_err(|err| {
-                        log::error!(target: LOG_TARGET, err:?; "pre_commit_sectors: invalid sealed_cid");
-                        Error::<T>::InvalidCid
-                    })?;
-
-                let deposit = calculate_pre_commit_deposit::<T>();
-
-                let entropy = owner.encode();
-                let randomness = get_randomness::<T>(
-                    DomainSeparationTag::SealRandomness,
-                    sector.seal_randomness_height,
-                    &entropy,
-                )?;
-
-                let sector_on_chain = SectorPreCommitOnChainInfo::new(
-                    sector.clone(),
-                    deposit,
-                    current_block,
-                    randomness,
-                );
-
-                // Push deal amounts for later verification
-                deal_amounts.try_push(sector_on_chain.info.deal_ids.len()).expect("Programmer error: cannot have more that MAX_SECTORS_PER_CALL deal_amount because of previous bounds");
-                // Push all unsealed_cids and deal amount to verify later.
-                unsealed_cids.try_push(commd.cid()).expect("Programmer error: cannot have more that MAX_SECTORS_PER_CALL unsealed_cids because of previous bounds");
-                // Push all deals to verify in one go later.
-                all_sector_deals.try_push((&sector_on_chain).into()).expect(
-                    "Programmer error: sector deals cannot be more that MAX_SECTORS_PER_CALL because of previous bounds",
-                );
-                // Add deposit to total deposit and push sector_on_chain to on_chain_sectors
-                // to avoid mutation of the SP for every sector.
-                total_deposit = total_deposit
-                    .checked_add(&deposit)
-                    .expect("Programmer error: Total deposit overflow should not happen because MAX_SECTORS_PER_CALL bound is lower than Balance::MAX");
-                on_chain_sectors
-                    .try_push(sector_on_chain)
-                    .expect("Programmer error: on chain sectors should fit in this BoundedVec due to previous validation");
-                // Push sector to BoundedVec for deposit event at the end
-                pre_committed_sectors
-                    .try_push(sector)
-                    .expect("Programmer error: sectors should fit in this BoundedVec due to previous validation");
-            }
-
-            let calculated_unsealed_cids =
-                T::Market::verify_deals_for_activation(&owner, all_sector_deals)?;
-            Self::check_commd_for_pre_commit(
-                calculated_unsealed_cids,
-                sector_amount,
-                unsealed_cids,
-                deal_amounts,
-            )?;
-
-            // Lock the pre-commit funds in the market account
-            T::Market::lock_pre_commit_funds(&owner, total_deposit)?;
-
-            StorageProviders::<T>::try_mutate(&owner, |maybe_sp| -> DispatchResult {
-                let sp = maybe_sp
-                    .as_mut()
-                    .ok_or(Error::<T>::StorageProviderNotFound)?;
-
-                // NOTE(@jmg-duarte,21/1/25): Not sure if this is still needed
-                sp.add_pre_commit_deposit(total_deposit)?;
-
-                for sector_on_chain in on_chain_sectors {
-                    sp.put_pre_committed_sector(sector_on_chain)
-                        .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-                }
-                Ok(())
-            })?;
-
-            Self::deposit_event(Event::SectorsPreCommitted {
-                block: current_block,
-                owner,
-                sectors: pre_committed_sectors,
-            });
-
-            Ok(())
+            crate::dispatchables::pre_commit_sectors::<T>(origin, sectors)
         }
 
         /// Allows the storage providers to submit proof for their pre-committed
@@ -570,131 +401,7 @@ pub mod pallet {
             origin: OriginFor<T>,
             sectors: BoundedVec<ProveCommitSector, ConstU32<MAX_SECTORS_PER_CALL>>,
         ) -> DispatchResult {
-            let owner = ensure_signed(origin)?;
-            let mut sp = StorageProviders::<T>::try_get(&owner)
-                .map_err(|_| Error::<T>::StorageProviderNotFound)?;
-            let current_block = <frame_system::Pallet<T>>::block_number();
-            // Create vectors for activating all prove commits at once.
-            let mut sector_deals = BoundedVec::new();
-            let mut new_sectors = BoundedVec::new();
-            let mut sector_numbers: BoundedVec<SectorNumber, ConstU32<MAX_SECTORS_PER_CALL>> =
-                BoundedVec::new();
-            let mut pre_commit_deposit_to_unlock = BalanceOf::<T>::zero();
-
-            for sector in sectors {
-                // Get pre-committed sector. This is the sector we are currently
-                // proving.
-                let precommit = sp
-                    .get_pre_committed_sector(sector.sector_number)
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-                let prove_commit_due =
-                    precommit.pre_commit_block_number + T::MaxProveCommitDuration::get();
-                ensure!(current_block < prove_commit_due, {
-                    log::error!(target: LOG_TARGET, "prove_commit_sectors: Prove commit submitted after the deadline. {current_block:?} > {prove_commit_due:?}");
-                    Error::<T>::ProveCommitAfterDeadline
-                });
-
-                // Validate the proof
-                validate_seal_proof::<T>(&owner, &precommit, sector.proofs)?;
-
-                // Sector deals that will be activated after the sector is
-                // successfully proven.
-                sector_deals
-                    .try_push(precommit.into())
-                    .expect("Programmer error: Sector deals should fit in bound of MAX_SECTORS");
-                sector_numbers
-                    .try_push(sector.sector_number)
-                    .expect("Programmer error: Sector numbers should fit in bound of MAX_SECTORS");
-                // Sector that will be activated and required to be periodically
-                // proven
-                let new_sector =
-                    SectorOnChainInfo::from_pre_commit(precommit.info.clone(), current_block);
-                new_sectors
-                    .try_push(new_sector)
-                    .expect("Programmer error: New sectors should fit in bound of MAX_SECTORS");
-
-                pre_commit_deposit_to_unlock += calculate_pre_commit_deposit::<T>();
-            }
-
-            // Activate the deals for the sectors that will be proven. This
-            // action is not applied if Err is returned from the extrinsic.
-            let compute_commd = sector_deals.len() > 0;
-            T::Market::activate_deals(&owner, sector_deals, compute_commd)?;
-
-            // Activate the new sectors and remove from pre-committed sectors.
-            sector_numbers.iter().zip(&new_sectors).try_for_each(
-                |(&sector_number, new_sector)| -> Result<(), Error<T>> {
-                    // Activate the new sector
-                    sp.activate_sector(sector_number, new_sector.clone())?;
-                    // Remove sector from the pre-committed map
-                    sp.remove_pre_committed_sector(sector_number)?;
-
-                    Ok(())
-                },
-            )?;
-
-            // Assign sectors to deadlines which specify when sectors needs
-            // to be proven
-            sp.assign_sectors_to_deadlines(
-                current_block,
-                new_sectors,
-                sp.info.window_post_partition_sectors,
-                T::MaxPartitionsPerDeadline::get(),
-                T::WPoStPeriodDeadlines::get(),
-                T::WPoStProvingPeriod::get(),
-                T::WPoStChallengeWindow::get(),
-                T::WPoStChallengeLookBack::get(),
-                T::FaultDeclarationCutoff::get(),
-            )
-            .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-            let sectors_proven = sector_numbers
-                .iter()
-                .map(|&sector_number| {
-                    // Find where the sector was placed. In worst case this goes through
-                    // all deadlines. It starts to look in the last partition of the
-                    // deadline. Usually the new sector will be there.
-                    let (deadline_idx, partition_number) = sp
-                        .deadlines
-                        .due
-                        .iter()
-                        .enumerate()
-                        .find_map(|(deadline_idx, deadline)| {
-                            deadline.partitions.iter().rev().find_map(
-                                |(partition_number, partition)| {
-                                    if partition.sectors.contains(&sector_number) {
-                                        Some((deadline_idx as u64, *partition_number))
-                                    } else {
-                                        None
-                                    }
-                                },
-                            )
-                        })
-                        .expect("sector should be assigned to a deadline");
-                    ProveCommitResult::new(sector_number, partition_number, deadline_idx)
-                })
-                .collect::<Vec<ProveCommitResult>>()
-                .try_into()
-                .expect("Programmer error: ProveCommitResult's should fit in bound of MAX_SECTORS");
-
-            // Reduce pre commit deposit amount in state
-            if let Some(pre_commit_deposits) = sp
-                .pre_commit_deposits
-                .checked_sub(&pre_commit_deposit_to_unlock)
-            {
-                sp.pre_commit_deposits = pre_commit_deposits
-            } else {
-                log::error!(target: LOG_TARGET, "catastrophe, failed to subtract from pre_commit_deposits {:?} - {:?} < 0", sp.pre_commit_deposits, pre_commit_deposit_to_unlock);
-                return Err(Error::<T>::FailedToReturnPreCommitDeposit.into());
-            };
-            // Unlock pre commit deposit funds.
-            T::Market::unlock_pre_commit_funds(&owner, pre_commit_deposit_to_unlock)?;
-            StorageProviders::<T>::set(owner.clone(), Some(sp));
-            Self::deposit_event(Event::SectorsProven {
-                owner,
-                sectors: sectors_proven,
-            });
-            Ok(())
+            crate::dispatchables::prove_commit_sectors::<T>(origin, sectors)
         }
 
         /// The SP uses this extrinsic to submit their Proof-of-Spacetime.
@@ -703,148 +410,7 @@ pub mod pallet {
             origin: OriginFor<T>,
             windowed_post: SubmitWindowedPoStParams,
         ) -> DispatchResult {
-            let owner = ensure_signed(origin)?;
-            let current_block = <frame_system::Pallet<T>>::block_number();
-            let mut sp = StorageProviders::<T>::try_get(&owner)
-                .map_err(|_| Error::<T>::StorageProviderNotFound)?;
-
-            for (idx, proof) in windowed_post.proofs.iter().enumerate() {
-                // Ensure proof matches the expected kind
-                ensure!(proof.post_proof == sp.info.window_post_proof_type, {
-                    log::error!(
-                        target: LOG_TARGET,
-                        "submit_window_post: idx: {}, expected PoSt type {:?} but received {:?} instead",
-                        idx,
-                        sp.info.window_post_proof_type,
-                        proof.post_proof
-                    );
-                    Error::<T>::InvalidProofType
-                });
-
-                ensure!(
-                    proof.proof_bytes.len() <= primitives::MAX_POST_PROOF_BYTES as usize,
-                    {
-                        log::error!("submit_window_post: invalid proof size");
-                        Error::<T>::PoStProofInvalid
-                    }
-                );
-            }
-
-            // If the proving period is in the future, we can't submit a proof yet
-            // Related issue: https://github.com/filecoin-project/specs-actors/issues/946
-            ensure!(current_block >= sp.proving_period_start, {
-                log::error!(target: LOG_TARGET,
-                    "proving period hasn't opened yet (current_block: {:?}, proving_period_start: {:?})",
-                    current_block,
-                    sp.proving_period_start
-                );
-                Error::<T>::InvalidDeadlineSubmission
-            });
-            let current_deadline = sp
-                .deadline_info(
-                    current_block,
-                    T::WPoStPeriodDeadlines::get(),
-                    T::WPoStProvingPeriod::get(),
-                    T::WPoStChallengeWindow::get(),
-                    T::WPoStChallengeLookBack::get(),
-                    T::FaultDeclarationCutoff::get(),
-                )
-                .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-            Self::validate_deadline(&current_deadline, &windowed_post)?;
-
-            // record sector as proven
-            let all_sectors = sp.sectors.clone();
-            let deadlines = sp.get_deadlines_mut();
-            deadlines
-                .record_proven(
-                    windowed_post.deadline as usize,
-                    &all_sectors,
-                    windowed_post.partitions.clone(),
-                )
-                .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-            // TODO(@th7nder,#592, 19/11/2024): handle faulty and recovered sectors, we don't take them into account now
-            let deadlines = &sp.deadlines;
-            let mut replicas = BoundedBTreeMap::new();
-            // Take all the sectors that were assigned to all of the partitions
-            for partition in &windowed_post.partitions {
-                // Deadline is validated by `Self::validate_deadline`, so we're sure it can be used as an index.
-                let deadline = &deadlines.due[windowed_post.deadline as usize];
-                let sectors = &deadline
-                    .partitions
-                    .get(&partition)
-                    .ok_or(Error::<T>::InvalidPartition)?
-                    .sectors;
-                for sector_number in sectors {
-                    // Sectors stored in the Storage Provider struct should be consistently stored, without breaking invariants.
-                    let sector_info = &sp.sectors[sector_number];
-                    let comm_r = Commitment::<CommR>::from_cid_bytes(&sector_info.sealed_cid)
-                        .expect("CommR to be validated on pre-commit");
-                    let _ = replicas
-                        .try_insert(
-                            *sector_number,
-                            PublicReplicaInfo {
-                                comm_r: comm_r.raw(),
-                            },
-                        )
-                        .map_err(|_| Error::<T>::TooManyReplicas)?;
-                }
-            }
-
-            log::debug!(target: LOG_TARGET, "submit_windowed_post: index {:?} challenge {:?}, replicas: {:?}",
-                current_deadline.idx,
-                current_deadline.challenge,
-                replicas
-            );
-
-            let entropy = owner.encode();
-            // The `chain_commit_epoch` should be `current_deadline.challenge` as per:
-            //
-            // These issues that were filed against the original implementation:
-            // * https://github.com/filecoin-project/specs-actors/issues/1094
-            // * https://github.com/filecoin-project/specs-actors/issues/1376
-            //
-            // The Go actors have this note:
-            // https://github.com/filecoin-project/specs-actors/blob/985cd0fa04578e262d68e0ef196f17df6f2434f2/actors/builtin/miner/miner_actor.go#L329-L332
-            //
-            // The fact that both Go and Rust actor implementations use the deadline challenge:
-            // * https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/miner/tests/miner_actor_test_wpost.rs#L99-L492
-            // * https://github.com/filecoin-project/specs-actors/blob/985cd0fa04578e262d68e0ef196f17df6f2434f2/actors/test/commit_post_test.go#L204-L215
-            // * https://github.com/filecoin-project/specs-actors/blob/985cd0fa04578e262d68e0ef196f17df6f2434f2/actors/test/terminate_sectors_scenario_test.go#L117-L128
-            //
-            // Further supported by the fact that Lotus and Curio (Lotus' replacement) don't use
-            // the ChainCommitEpoch variable from the SubmitWindowedPostParams
-            // * https://github.com/filecoin-project/lotus/blob/4f70204342ce83671a7a261147a18865f1618967/storage/wdpost/wdpost_run.go#L334-L338
-            // * https://github.com/filecoin-project/lotus/blob/4f70204342ce83671a7a261147a18865f1618967/curiosrc/window/compute_do.go#L68-L72
-            // * https://github.com/filecoin-project/curio/blob/45373f7fc0431e41f987ad348df7ae6e67beaff9/tasks/window/compute_do.go#L71-L75
-            let randomness = get_randomness::<T>(
-                DomainSeparationTag::WindowedPoStChallengeSeed,
-                current_deadline.challenge,
-                &entropy,
-            )?;
-
-            let mut proofs = BoundedVec::new();
-            for proof in windowed_post.proofs {
-                proofs
-                    .try_push(proof.proof_bytes)
-                    .map_err(|_| Error::<T>::TooManyProofs)?;
-            }
-
-            T::ProofVerification::verify_post(
-                sp.info.window_post_proof_type,
-                randomness,
-                replicas,
-                proofs,
-            )?;
-
-            log::debug!(target: LOG_TARGET, "submit_windowed_post: proof recorded");
-
-            // Store new storage provider state
-            StorageProviders::<T>::set(owner.clone(), Some(sp));
-            Self::deposit_event(Event::ValidPoStSubmitted { owner });
-
-            Ok(())
+            crate::dispatchables::submit_windowed_post::<T>(origin, windowed_post)
         }
 
         /// The SP uses this extrinsic to declare some sectors as faulty. Letting the system know it will not submit PoSt for the next deadline.
@@ -853,83 +419,7 @@ pub mod pallet {
         /// * <https://github.com/filecoin-project/builtin-actors/blob/82d02e58f9ef456aeaf2a6c737562ac97b22b244/actors/miner/src/lib.rs#L2648>
         #[pallet::call_index(4)]
         pub fn declare_faults(origin: OriginFor<T>, params: DeclareFaultsParams) -> DispatchResult {
-            let owner = ensure_signed(origin)?;
-            let current_block = <frame_system::Pallet<T>>::block_number();
-            let mut sp = StorageProviders::<T>::try_get(&owner)
-                .map_err(|_| Error::<T>::StorageProviderNotFound)?;
-
-            let mut to_process = DeadlineSectorMap::new();
-            for term in &params.faults {
-                let deadline = term.deadline;
-                let partition = term.partition;
-
-                // Check if the sectors passed are empty
-                if term.sectors.is_empty() {
-                    log::error!(target: LOG_TARGET, "declare_faults: [deadline: {}, partition: {}] cannot add empty sectors", deadline, partition);
-                    return Err(Error::<T>::GeneralPalletError(
-                        crate::error::GeneralPalletError::DeadlineErrorCouldNotAddSectors,
-                    )
-                    .into());
-                }
-
-                to_process
-                    .try_insert(deadline, partition, term.sectors.clone())
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-            }
-
-            for (&deadline_idx, partition_map) in to_process.into_iter() {
-                log::debug!(target: LOG_TARGET, "declare_faults: Processing deadline index: {deadline_idx}");
-                // Check deadline index to avoid doing any work if it is wrong.
-                ensure!(
-                    (deadline_idx as usize) < sp.deadlines.due.len(),
-                    Error::<T>::GeneralPalletError(
-                        crate::error::GeneralPalletError::DeadlineErrorDeadlineIndexOutOfRange
-                    )
-                );
-                // Get the target deadline
-                // We're deviating from the original implementation by using the `sp.proving_period_start`
-                // instead of calculating it here, but we couldn't find a reason to do it in another way
-                //
-                // References:
-                // * https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/miner/src/lib.rs#L2436-L2449
-                // * https://github.com/eigerco/polka-storage/pull/192#discussion_r1715067288
-                let target_dl = DeadlineInfo::new(
-                    current_block,
-                    sp.proving_period_start,
-                    deadline_idx,
-                    T::WPoStPeriodDeadlines::get(),
-                    T::WPoStProvingPeriod::get(),
-                    T::WPoStChallengeWindow::get(),
-                    T::WPoStChallengeLookBack::get(),
-                    T::FaultDeclarationCutoff::get(),
-                )
-                .and_then(DeadlineInfo::next_not_elapsed)
-                .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-                // https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/miner/src/lib.rs#L2451-L2458
-                ensure!(!target_dl.fault_cutoff_passed(), {
-                    log::error!(target: LOG_TARGET, "declare_faults: Late fault declaration at deadline {:?}. {:?} >= {:?}", deadline_idx, current_block, target_dl.fault_cutoff);
-                    Error::<T>::FaultDeclarationTooLate
-                });
-
-                let fault_expiration_block = target_dl.last() + T::FaultMaxAge::get();
-                log::debug!(target: LOG_TARGET, "declare_faults: Getting deadline[{deadline_idx}]");
-                let dl = sp
-                    .deadlines
-                    .load_deadline_mut(deadline_idx as usize)
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-                dl.record_faults(&sp.sectors, partition_map, fault_expiration_block)
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-            }
-
-            StorageProviders::<T>::set(owner.clone(), Some(sp));
-            Self::deposit_event(Event::FaultsDeclared {
-                owner,
-                faults: params.faults,
-            });
-
-            Ok(())
+            crate::dispatchables::declare_faults::<T>(origin, params)
         }
 
         /// This extrinsic allows an SP to declare some faulty sectors as recovering.
@@ -943,73 +433,7 @@ pub mod pallet {
             origin: OriginFor<T>,
             params: DeclareFaultsRecoveredParams,
         ) -> DispatchResult {
-            let owner = ensure_signed(origin)?;
-            let current_block = <frame_system::Pallet<T>>::block_number();
-            let mut sp = StorageProviders::<T>::try_get(&owner)
-                .map_err(|_| Error::<T>::StorageProviderNotFound)?;
-            let mut to_process = DeadlineSectorMap::new();
-
-            for term in &params.recoveries {
-                let deadline = term.deadline;
-                let partition = term.partition;
-
-                // Check if the sectors passed are empty
-                if term.sectors.is_empty() {
-                    log::error!(target: LOG_TARGET, "declare_faults_recovered: sectors cannot be empty for deadline: {:?}, partition: {:?}", deadline, partition);
-                    return Err(Error::<T>::GeneralPalletError(
-                        crate::error::GeneralPalletError::DeadlineErrorCouldNotAddSectors,
-                    )
-                    .into());
-                }
-
-                to_process
-                    .try_insert(deadline, partition, term.sectors.clone())
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-            }
-
-            for (&deadline_idx, partition_map) in to_process.0.iter() {
-                log::debug!(target: LOG_TARGET, "declare_faults_recovered: processing deadline index: {deadline_idx}");
-                // Check deadline index to avoid doing any work if it is wrong.
-                ensure!(
-                    (deadline_idx as usize) < sp.deadlines.due.len(),
-                    Error::<T>::GeneralPalletError(
-                        crate::error::GeneralPalletError::DeadlineErrorDeadlineIndexOutOfRange
-                    )
-                );
-                // Get the deadline
-                let target_dl = DeadlineInfo::new(
-                    current_block,
-                    sp.proving_period_start,
-                    deadline_idx,
-                    T::WPoStPeriodDeadlines::get(),
-                    T::WPoStProvingPeriod::get(),
-                    T::WPoStChallengeWindow::get(),
-                    T::WPoStChallengeLookBack::get(),
-                    T::FaultDeclarationCutoff::get(),
-                )
-                .and_then(DeadlineInfo::next_not_elapsed)
-                .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-                ensure!(!target_dl.fault_cutoff_passed(), {
-                    log::error!(target: LOG_TARGET, "declare_faults: late fault declaration at deadline {:?}. {:?} >= {:?}",
-                        deadline_idx, current_block, target_dl.fault_cutoff);
-                    Error::<T>::FaultRecoveryTooLate
-                });
-                let dl = sp
-                    .deadlines
-                    .load_deadline_mut(deadline_idx as usize)
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-                dl.declare_faults_recovered(&sp.sectors, partition_map)
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-            }
-
-            StorageProviders::<T>::insert(owner.clone(), sp);
-            Self::deposit_event(Event::FaultsRecovered {
-                owner,
-                recoveries: params.recoveries,
-            });
-
-            Ok(())
+            crate::dispatchables::declare_faults_recovered::<T>(origin, params)
         }
 
         /// Marks some sectors as terminated at the present block, earlier than their
@@ -1022,69 +446,7 @@ pub mod pallet {
             origin: OriginFor<T>,
             params: TerminateSectorsParams,
         ) -> DispatchResult {
-            let owner = ensure_signed(origin)?;
-            let current_block = <frame_system::Pallet<T>>::block_number();
-            let mut sp = StorageProviders::<T>::try_get(&owner)
-                .map_err(|_| Error::<T>::StorageProviderNotFound)?;
-
-            let mut to_process = DeadlineSectorMap::new();
-
-            for term in params.terminations.iter() {
-                let deadline = term.deadline;
-                let partition = term.partition;
-
-                to_process
-                    .try_insert(deadline, partition, term.sectors.clone())
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-            }
-
-            let sectors = sp
-                .sectors
-                .iter()
-                .map(|(_sector_number, info)| info)
-                .cloned()
-                .collect::<Vec<_>>();
-            for (&deadline_idx, partition_sectors) in to_process.into_iter() {
-                ensure!(
-                    deadline_is_mutable(
-                        sp.proving_period_start,
-                        deadline_idx,
-                        current_block,
-                        T::WPoStPeriodDeadlines::get(),
-                        T::WPoStProvingPeriod::get(),
-                        T::WPoStChallengeWindow::get(),
-                        T::WPoStChallengeLookBack::get(),
-                        T::FaultDeclarationCutoff::get(),
-                    )
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?,
-                    {
-                        log::error!(target: LOG_TARGET, "cannot terminate sectors in immutable deadline {}", deadline_idx);
-                        Error::<T>::CannotTerminateImmutableDeadline
-                    }
-                );
-
-                let deadline = sp
-                    .deadlines
-                    .load_deadline_mut(deadline_idx as usize)
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-                deadline
-                    .terminate_sectors(current_block, &sectors, partition_sectors)
-                    .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-                sp.early_terminations.insert(deadline_idx);
-            }
-
-            // Update storage provider state
-            StorageProviders::<T>::insert(&owner, sp);
-
-            Self::process_early_terminations(current_block, &owner)?;
-
-            Self::deposit_event(Event::SectorsTerminated {
-                owner,
-                terminations: params.terminations,
-            });
-            Ok(())
+            crate::dispatchables::terminate_sectors::<T>(origin, params)
         }
     }
 
@@ -1096,8 +458,8 @@ pub mod pallet {
         }
 
         fn on_finalize(current_block: BlockNumberFor<T>) {
-            Self::check_precommited_sectors(current_block);
-            Self::check_deadlines(current_block);
+            crate::hooks::check_precommited_sectors::<T>(current_block);
+            crate::hooks::check_deadlines::<T>(current_block);
         }
     }
 
@@ -1108,37 +470,6 @@ pub mod pallet {
     }
 
     impl<T: Config> Pallet<T> {
-        /// Gets the current deadline of the storage provider.
-        ///
-        /// If there is no Storage Provider of given AccountId returns [`Option::None`].
-        /// May exceptionally return [`Option::None`] when
-        /// conversion between BlockNumbers fails, but technically should never happen.
-        pub fn current_deadline(
-            storage_provider: &T::AccountId,
-        ) -> Option<ExternalDeadlineInfo<BlockNumberFor<T>>> {
-            let sp = StorageProviders::<T>::try_get(storage_provider).ok()?;
-            let current_block = <frame_system::Pallet<T>>::block_number();
-
-            let deadline = sp
-                .deadline_info(
-                    current_block,
-                    T::WPoStPeriodDeadlines::get(),
-                    T::WPoStProvingPeriod::get(),
-                    T::WPoStChallengeWindow::get(),
-                    T::WPoStChallengeLookBack::get(),
-                    T::FaultDeclarationCutoff::get(),
-                )
-                .ok()?;
-
-            Some(ExternalDeadlineInfo {
-                deadline_index: deadline.idx,
-                open: deadline.is_open(),
-                challenge_block: deadline.challenge,
-                start: deadline.open_at,
-                close: deadline.close_at,
-            })
-        }
-
         /// Gets the next, not yet opened, deadline of the storage provider.
         ///
         /// If there is no Storage Provider of given AccountId returns [`Option::None`].
@@ -1206,524 +537,5 @@ pub mod pallet {
 
             Some(primitives::pallets::DeadlineState { partitions })
         }
-
-        fn validate_expiration(
-            curr_block: BlockNumberFor<T>,
-            activation: BlockNumberFor<T>,
-            expiration: BlockNumberFor<T>,
-        ) -> Result<(), Error<T>> {
-            log::debug!(target: LOG_TARGET, "validate_expiration: {:?} {:?} {:?}", curr_block, activation, expiration);
-            // Expiration must be after activation. Check this explicitly to avoid an underflow below.
-            ensure!(
-                expiration >= activation,
-                Error::<T>::ExpirationBeforeActivation
-            );
-            // expiration cannot be less than minimum after activation
-            ensure!(
-                expiration - activation >= T::MinSectorExpiration::get(),
-                Error::<T>::ExpirationTooSoon
-            );
-            // expiration cannot exceed MaxSectorExpiration from now
-            ensure!(
-                expiration < curr_block + T::MaxSectorExpiration::get(),
-                Error::<T>::ExpirationTooLong,
-            );
-            // total sector lifetime cannot exceed SectorMaximumLifetime for the sector's seal proof
-            ensure!(
-                expiration - activation < T::SectorMaximumLifetime::get(),
-                Error::<T>::MaxSectorLifetimeExceeded
-            );
-            Ok(())
-        }
-
-        /// Check whether the given deadline is valid for PoSt submission.
-        ///
-        /// Fails if:
-        /// - The given deadline is not open.
-        /// - There is and deadline index mismatch.
-        /// - The block the deadline was committed at is after challenge height.
-        ///
-        /// Reference:
-        /// * <https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/miner/src/lib.rs#L591-L626>
-        fn validate_deadline(
-            current_deadline: &DeadlineInfo<BlockNumberFor<T>>,
-            post_params: &SubmitWindowedPoStParams,
-        ) -> Result<(), Error<T>> {
-            // Ensure the deadline is open
-            ensure!(current_deadline.is_open(), {
-                log::error!(target: LOG_TARGET, "validate_deadline: {current_deadline:?}, deadline isn't open");
-                Error::<T>::InvalidDeadlineSubmission
-            });
-
-            // Ensure the deadline index matches the one in the post params
-            ensure!(post_params.deadline == current_deadline.idx, {
-                log::error!(target: LOG_TARGET, "validate_deadline: given index does not match current index {} != {}", post_params.deadline, current_deadline.idx);
-                Error::<T>::InvalidDeadlineSubmission
-            });
-
-            Ok(())
-        }
-
-        /// Goes through all of the registered storage providers and checks if they have any expired pre committed sectors.
-        /// If there are any sectors that are expired the total deposit amount for all those sectors will be slashed.
-        ///
-        /// References:
-        /// * <https://github.com/filecoin-project/builtin-actors/blob/82d02e58f9ef456aeaf2a6c737562ac97b22b244/actors/miner/src/state.rs#L1071>
-        /// * <https://github.com/filecoin-project/builtin-actors/blob/82d02e58f9ef456aeaf2a6c737562ac97b22b244/actors/miner/src/state.rs#L1054>
-        fn check_precommited_sectors(current_block: BlockNumberFor<T>) {
-            const LOG_TARGET: &'static str = "runtime::storage_provider::check_precommited_sectors";
-
-            // TODO(@th7nder,31/07/2024): this approach is suboptimal, as it's time complexity is O(StorageProviders * PreCommitedSectors).
-            // We can reduce this by indexing pre-committed sectors by BlockNumber in which they're supposed to be activated in PreCommit and remove them in ProveCommit.
-            log::debug!(target: LOG_TARGET, "checking pre_commited_sectors for block: {:?}", current_block);
-
-            // We cannot modify storage map while inside `iter_keys()` as docs say it's undefined results.
-            // And we can use `alloc::Vec`, because it's bounded by StorageProviders data structure anyways.
-            let storage_providers: Vec<_> = StorageProviders::<T>::iter_keys().collect();
-            for storage_provider in storage_providers {
-                log::debug!(target: LOG_TARGET, "checking storage provider {:?}", storage_provider);
-                let Ok(mut state) = StorageProviders::<T>::try_get(storage_provider.clone()) else {
-                    log::error!(target: LOG_TARGET, "catastrophe, couldn't find a storage provider based on key. it should have been there...");
-                    continue;
-                };
-
-                let (expired, slash_amount) =
-                    Self::detect_expired_precommit_sectors(current_block, &state);
-                if expired.is_empty() {
-                    return;
-                }
-
-                let mut removed_sectors = BoundedVec::new();
-                log::info!(target: LOG_TARGET, "found {} expired pre committed sectors for {:?}", expired.len(), storage_provider);
-                for sector_number in expired {
-                    // Expired sectors should be removed, because in other case they'd be processed twice in the next block.
-                    if let Ok(()) = state.remove_pre_committed_sector(sector_number) {
-                        removed_sectors.force_push(sector_number)
-                    } else {
-                        log::error!(target: LOG_TARGET, "catastrophe, failed to remove sector {} for {:?}", sector_number, storage_provider);
-                        continue;
-                    };
-                }
-
-                let Some(slashed_deposits) = state.pre_commit_deposits.checked_sub(&slash_amount)
-                else {
-                    log::error!(target: LOG_TARGET, "catastrophe, failed to subtract from pre_commit_deposits {:?} - {:?} < 0", state.pre_commit_deposits, slash_amount);
-                    continue;
-                };
-                state.pre_commit_deposits = slashed_deposits;
-
-                // PRE-COND: currency was previously reserved in pre_commit
-                let Ok(()) = T::Market::slash_pre_commit_funds(&storage_provider, slash_amount)
-                else {
-                    log::error!(target: LOG_TARGET, "failed to slash.. amount: {:?}, storage_provider: {:?}", slash_amount, storage_provider);
-                    continue;
-                };
-
-                StorageProviders::<T>::insert(&storage_provider, state);
-                Self::deposit_event(Event::<T>::SectorsSlashed {
-                    owner: storage_provider,
-                    sector_numbers: removed_sectors,
-                })
-            }
-        }
-
-        /// Checks whether pre-committed sectors are expired and calculates slash amount.
-        ///
-        /// THIS FUNCTION DOES NOT HANDLE ERRORS!
-        /// Code in hooks is assumed infallible and operates under invariants.
-        ///
-        /// Returns an array of expired sector numbers and the total deposit to be slashed.
-        fn detect_expired_precommit_sectors(
-            curr_block: BlockNumberFor<T>,
-            state: &StorageProviderState<T::PeerId, BalanceOf<T>, BlockNumberFor<T>>,
-        ) -> (
-            BoundedVec<SectorNumber, ConstU32<MAX_SECTORS>>,
-            BalanceOf<T>,
-        ) {
-            let mut expired_sectors: BoundedVec<SectorNumber, ConstU32<MAX_SECTORS>> =
-                BoundedVec::new();
-            let mut to_be_slashed = BalanceOf::<T>::zero();
-
-            for (sector_number, sector) in &state.pre_committed_sectors {
-                // Expiration marks the time for a block when it was supposed to be proven by `prove_commit` ultimately.
-                // If it's still in `pre_commited_sectors` and `curr_block` is past this time, it means it was not.
-                if curr_block >= sector.info.expiration {
-                    let Ok(()) = expired_sectors.try_push(*sector_number) else {
-                        log::error!(target: LOG_TARGET, "detect_expired_precommit_sectors: invariant violated, expired_sectors bounded_vec's capacity < state.pre_committed_sectors capacity, sector: {}", sector_number);
-                        continue;
-                    };
-                    let Some(result) = to_be_slashed.checked_add(&sector.pre_commit_deposit) else {
-                        log::error!(target: LOG_TARGET, "detect_expired_precommit_sectors: invariant violated, overflow in adding slash deposit: sector: {}, current: {:?}, to add: {:?}", sector_number, to_be_slashed, sector.pre_commit_deposit);
-                        continue;
-                    };
-                    to_be_slashed = result;
-                }
-            }
-
-            (expired_sectors, to_be_slashed)
-        }
-
-        /// Goes through each Storage Provider and its current deadline.
-        ///
-        /// If the deadline elapsed (current_block >= deadline.close_at) it checks all of the partitions and their sectors.
-        /// If a proof for a partition has not been submitted, all sectors in the partition are marked as faulty.
-        /// A deadline is checked once every [`T::WPoStProvingPeriod`]. If a Partition was marked as faulty in a deadline (deadline_idx, proving_period_idx),
-        /// it's rechecked in the next [`T::WPoStProvingPeriod`] in the next deadline (deadline_idx, proving_period_idx + 1).
-        /// `pre_commit_deposit` is slashed by 1 for each partition for each proving period a partition is faulty.
-        ///
-        /// TODO:
-        /// - If a partition is faulty for too long [`T::FaultMaxAge`], it needs to be be terminated. (#165, #167)
-        /// - A proper slashing mechanism `pre_commit_deposit` and calculation. (#187)
-        ///
-        /// Reference implementation:
-        /// * <https://github.com/filecoin-project/builtin-actors/blob/82d02e58f9ef456aeaf2a6c737562ac97b22b244/actors/miner/src/state.rs#L1128>
-        /// * <https://github.com/filecoin-project/builtin-actors/blob/82d02e58f9ef456aeaf2a6c737562ac97b22b244/actors/miner/src/state.rs#L1192>
-        fn check_deadlines(current_block: BlockNumberFor<T>) {
-            const LOG_TARGET: &'static str = "runtime::storage_provider::check_deadlines";
-            log::debug!(target: LOG_TARGET, "block: {:?}", current_block);
-
-            // We cannot modify storage map while inside `iter_keys()` as docs say it's undefined results.
-            // And we can use `alloc::Vec`, because it's bounded by StorageProviders data structure anyways.
-            let storage_providers: Vec<_> = StorageProviders::<T>::iter_keys().collect();
-            // TODO(@th7nder,13/08/2024): this approach is suboptimal, as it's time complexity is O(StorageProviders * PreCommitedSectors).
-            // We can reduce this by indexing pre-committed sectors by BlockNumber in which they're supposed to be activated in PreCommit and remove them in ProveCommit.
-            for storage_provider in storage_providers {
-                log::debug!(target: LOG_TARGET, "block: {:?}, checking storage provider {:?}", current_block, storage_provider);
-                let Ok(mut state) = StorageProviders::<T>::try_get(storage_provider.clone()) else {
-                    log::error!(target: LOG_TARGET, "missing storage provider {:?} (should have been added before)", storage_provider);
-                    continue;
-                };
-
-                if current_block < state.proving_period_start {
-                    log::debug!(target: LOG_TARGET, "skipping checking sp: {:?} on block: {:?} < proving_start {:?}, because it hasn't started yet.",
-                    storage_provider, current_block, state.proving_period_start);
-                    continue;
-                }
-
-                let Ok(current_deadline) = state.deadline_info(
-                    current_block,
-                    T::WPoStPeriodDeadlines::get(),
-                    T::WPoStProvingPeriod::get(),
-                    T::WPoStChallengeWindow::get(),
-                    T::WPoStChallengeLookBack::get(),
-                    T::FaultDeclarationCutoff::get(),
-                ) else {
-                    log::error!(target: LOG_TARGET, "block: {:?}, there are no deadlines for storage provider {:?}", current_block, storage_provider);
-                    continue;
-                };
-
-                if !current_deadline.period_started() {
-                    log::debug!(target: LOG_TARGET, "block: {:?}, period for deadline {:?}, sp {:?} has not yet started...", current_block, current_deadline.idx, storage_provider);
-                    continue;
-                }
-
-                if !current_deadline.has_elapsed() {
-                    log::debug!(target: LOG_TARGET,
-                    "block: {:?}, deadline {:?} for sp {:?} not yet elapsed. open_at: {:?} < current {:?} < close_at {:?}",
-                    current_block,
-                    current_deadline.idx, storage_provider, current_deadline.open_at, current_block, current_deadline.close_at
-                    );
-                    continue;
-                }
-
-                log::debug!(target: LOG_TARGET, "block: {:?}, checking storage provider {:?} deadline: {:?}",
-                    current_block,
-                    storage_provider,
-                    current_deadline.idx,
-                );
-
-                let Ok(deadline) =
-                    (&mut state.deadlines).load_deadline_mut(current_deadline.idx as usize)
-                else {
-                    log::error!(target: LOG_TARGET, "block: {:?}, failed to get deadline {}, sp: {:?}",
-                        current_block, current_deadline.idx, storage_provider);
-                    continue;
-                };
-
-                let mut faulty_partitions_amount = 0;
-                // Create collection for fault partitions, 1 event per SP
-                let mut faulty_partitions: BTreeMap<
-                    PartitionNumber,
-                    BoundedBTreeSet<SectorNumber, ConstU32<MAX_SECTORS>>,
-                > = BTreeMap::new();
-                for (partition_number, partition) in deadline.partitions.iter_mut() {
-                    if partition.sectors.len() == 0 {
-                        continue;
-                    }
-                    // WindowPoSt Proof was submitted for a partition.
-                    if deadline.partitions_posted.contains(&partition_number) {
-                        continue;
-                    }
-
-                    log::debug!(target: LOG_TARGET, "block: {:?}, going through partition: {:?}", current_block, partition);
-
-                    // Mark all Sectors in a partition as faulty
-                    let fault_expiration_block = current_deadline.last() + T::FaultMaxAge::get();
-                    let Ok(new_faults) = partition.record_faults(
-                        &state.sectors,
-                        &partition.sectors.clone(),
-                        fault_expiration_block,
-                    ) else {
-                        log::error!(target: LOG_TARGET, "block: {:?}, failed to mark {} sectors as faulty, deadline: {}, sp: {:?}",
-                            current_block, partition.sectors.len(), current_deadline.idx, storage_provider);
-                        continue;
-                    };
-
-                    if let Err(e) =
-                        Self::process_early_terminations(current_block, &storage_provider)
-                    {
-                        log::error!(target: LOG_TARGET, "could not process early terminations for {storage_provider:?}: {e:?}");
-                        continue;
-                    }
-                    log::info!(target: LOG_TARGET, "block: {:?}, sp: {:?}, detected partition {} with {} new faults...",
-                    current_block, storage_provider, partition_number, new_faults.len());
-
-                    if new_faults.len() > 0 {
-                        faulty_partitions.insert(*partition_number, new_faults.try_into().expect("should be able to create BoundedBTreeSet due to input being bounded"));
-                        faulty_partitions_amount += 1;
-                    }
-                }
-
-                // TODO(@th7nder,[#106,#187],08/08/2024): figure out slashing amounts (for continued faults, new faults).
-                if faulty_partitions_amount > 0 {
-                    log::warn!(target: LOG_TARGET, "block: {:?}, sp: {:?}, deadline: {:?} - should have slashed {} partitions...",
-                        current_block,
-                        storage_provider,
-                        current_deadline.idx,
-                        faulty_partitions_amount,
-                    );
-                    Self::deposit_event(Event::PartitionsFaulty {
-                        owner: storage_provider.clone(),
-                        faulty_partitions: faulty_partitions.try_into().expect("should be able to create a BTreeMap with a MAX_PARTITIONS_PER_DEADLINE bound after iterating over a map with the same bound"),
-                    })
-                } else if !deadline.partitions.is_empty() {
-                    log::info!(target: LOG_TARGET, "block: {:?}, sp: {:?}, deadline: {:?} - all proofs submitted on time.",
-                        current_block,
-                        storage_provider,
-                        current_deadline.idx,
-                    );
-                }
-
-                // Reset posted partitions, as deadline has been processed.
-                // Next processing will happen in the next proving period.
-                deadline.partitions_posted = BoundedBTreeSet::new();
-                state
-                    .advance_deadline(
-                        current_block,
-                        T::WPoStPeriodDeadlines::get(),
-                        T::WPoStProvingPeriod::get(),
-                        T::WPoStChallengeWindow::get(),
-                        T::WPoStChallengeLookBack::get(),
-                        T::FaultDeclarationCutoff::get(),
-                    )
-                    .expect("Could not advance deadline");
-                StorageProviders::<T>::insert(storage_provider, state);
-            }
-        }
-
-        /// Verifies that the unsealed_cid (CommD) and checks that it matches the given unsealed CID.
-        fn check_commd_for_pre_commit(
-            calculated_unsealed_cid: BoundedVec<Option<Cid>, ConstU32<MAX_DEALS_PER_SECTOR>>,
-            sector_amount: usize,
-            unsealed_cids: BoundedVec<Cid, ConstU32<MAX_DEALS_PER_SECTOR>>,
-            deal_amounts: BoundedVec<usize, ConstU32<MAX_DEALS_PER_SECTOR>>,
-        ) -> Result<(), Error<T>> {
-            ensure!(calculated_unsealed_cid.len() == sector_amount, {
-                log::error!(target: LOG_TARGET, "check_commd_for_pre_commit: failed to verify deals, invalid calculated_commd length: {}", calculated_unsealed_cid.len());
-                Error::<T>::CouldNotVerifySectorForPreCommit
-            });
-
-            for (i, unsealed_cid) in unsealed_cids.into_iter().enumerate() {
-                if deal_amounts[i] > 0 {
-                    let Some(calculated_commd) = calculated_unsealed_cid[i] else {
-                        log::error!(target: LOG_TARGET, "check_commd_for_pre_commit: commd for the deals at index {i} from verify_deals is None...");
-                        fail!(Error::<T>::CouldNotVerifySectorForPreCommit)
-                    };
-
-                    ensure!(calculated_commd == unsealed_cid, {
-                        log::error!(target: LOG_TARGET, "check_commd_for_pre_commit: calculated_commd at index {i} != sector.unsealed_cid, {:?} != {:?}", calculated_commd, unsealed_cid);
-                        Error::<T>::InvalidUnsealedCidForSector
-                    });
-                }
-            }
-            Ok(())
-        }
-
-        /// Checks if the sectors submitted for pre-commit by the SP are valid.
-        /// Checks are
-        /// - Sector number limit (cannot be higher than MAX_SECTORS)
-        /// - The proof type must correspond to the proof type submitted during registration.
-        /// - The sector number must not be used previously.
-        fn validate_sector_for_pre_commit(
-            sp: &StorageProviderState<T::PeerId, BalanceOf<T>, BlockNumberFor<T>>,
-            sector: &SectorPreCommitInfo<BlockNumberFor<T>>,
-        ) -> Result<(), Error<T>> {
-            let sector_number = sector.sector_number;
-
-            ensure!(
-                sp.info.window_post_proof_type == sector.seal_proof.registered_window_post_proof(),
-                Error::<T>::InvalidProofType
-            );
-            ensure!(
-                !sp.pre_committed_sectors.contains_key(&sector_number)
-                    && !sp.sectors.contains_key(&sector_number),
-                Error::<T>::SectorNumberAlreadyUsed
-            );
-            Ok(())
-        }
-
-        /// Processes terminations for the given account (should be a registered SP).
-        /// Clears all early terminations and calls `on_sectors_terminate` when finished.
-        fn process_early_terminations(
-            current_block: BlockNumberFor<T>,
-            owner: &T::AccountId,
-        ) -> Result<(), Error<T>> {
-            let mut state = StorageProviders::<T>::try_get(owner)
-                .map_err(|_| Error::<T>::StorageProviderNotFound)?;
-            let result = state
-                .pop_early_terminations(
-                    T::AddressedPartitionsMax::get(),
-                    T::AddressedSectorsMax::get(),
-                )
-                .map_err(|e| Error::<T>::GeneralPalletError(e))?;
-
-            // Nothing to do, don't waste any time.
-            // This can happen if we end up processing early terminations
-            // before the cron callback fires.
-            if result.is_empty() {
-                log::info!(target: LOG_TARGET, "no early terminations");
-                return Ok(());
-            }
-
-            let mut sectors_with_data = Vec::new();
-            // Check whether sectors have expired, if not push to sectors_with_data for later processing.
-            // Process in Market pallet `on_sectors_terminate`.
-            // TODO(@aidan46, #452, 2024-10-14): Figure out economics to apply early termination penalty.
-            for (&expiry, sector_numbers) in result.sectors.iter() {
-                for sector_number in sector_numbers {
-                    // I am not 100% sure this is correct. In FC they use deal weight to determine.
-                    // Deal weight is a function of space times the duration of a deal.
-                    if expiry < current_block {
-                        sectors_with_data.push(*sector_number);
-                    }
-                }
-            }
-
-            // Terminate deals
-            let terminated_data = sectors_with_data.try_into().expect(
-                "The sectors in the result can never be more than MAX_DEALS_PER_SECTOR due to previous bounds, this should not fail",
-            );
-
-            T::Market::on_sectors_terminate(owner, terminated_data)
-                .map_err(|_| Error::<T>::CouldNotTerminateDeals)?;
-
-            // Update storage provider state
-            StorageProviders::<T>::insert(owner, state);
-
-            Ok(())
-        }
-    }
-
-    /// Calculate the required pre commit deposit amount
-    fn calculate_pre_commit_deposit<T: Config>() -> BalanceOf<T> {
-        BalanceOf::<T>::one() // TODO(@aidan46, #106, 2024-06-24): Set a logical value or calculation
-    }
-
-    fn validate_seal_proof<T: Config>(
-        owner: &T::AccountId,
-        precommit: &SectorPreCommitOnChainInfo<BalanceOf<T>, BlockNumberFor<T>>,
-        proofs: BoundedVec<
-            BoundedVec<u8, ConstU32<MAX_SEAL_PROOF_BYTES>>,
-            ConstU32<MAX_POREP_PROOFS_PER_BLOCK>,
-        >,
-    ) -> Result<(), DispatchError> {
-        let max_proof_size = precommit.info.seal_proof.proof_size();
-
-        // Check proof size
-        if let Some(proof) = proofs.iter().filter(|p| p.len() > max_proof_size).nth(0) {
-            log::error!(target: LOG_TARGET, "sector proof size {} exceeds max {}", proof.len(), max_proof_size);
-            return Err(Error::<T>::InvalidProof)?;
-        }
-
-        let current_block_number = <frame_system::Pallet<T>>::block_number();
-
-        // https://github.com/filecoin-project/builtin-actors/blob/a45fb87910bca74d62215b0d58ed90cf78b6c8ff/actors/miner/src/lib.rs#L4865
-        // Check if we are too early with the proof submit
-        let interactive_block_number =
-            precommit.pre_commit_block_number + T::PreCommitChallengeDelay::get();
-
-        if current_block_number < interactive_block_number {
-            log::error!(target: LOG_TARGET, "too early to prove sector: current_block_number: {current_block_number:?} < interactive_block_number: {interactive_block_number:?}");
-            return Err(Error::<T>::InvalidProof)?;
-        }
-
-        // Validate the data commitment
-        let commd = Commitment::<CommD>::from_cid_bytes(&precommit.info.unsealed_cid[..])
-            .map_err(|err| {
-                log::error!(target: LOG_TARGET, err:?; "validate_seal_proof: invalid unsealed_cid {:?}", &precommit.info.unsealed_cid);
-                Error::<T>::InvalidCid
-            })?;
-
-        // Validate the replica commitment
-        let commr = Commitment::<CommR>::from_cid_bytes(&precommit.info.sealed_cid[..])
-            .map_err(|err| {
-                log::error!(target: LOG_TARGET, err:?; "validate_seal_proof: invalid sealed_cid {:?}", &precommit.info.sealed_cid);
-                Error::<T>::InvalidCid
-            })?;
-
-        let entropy = owner.encode();
-        let interactive_randomness = get_randomness::<T>(
-            DomainSeparationTag::InteractiveSealChallengeSeed,
-            interactive_block_number,
-            &entropy,
-        )?;
-
-        let prover_id = derive_prover_id(owner);
-
-        log::debug!(target: LOG_TARGET, "Performing prove commit for, seal_randomness_height {:?}, pre_commit_block: {:?}, prove_commit_block: {:?}, entropy: {}, ticket: {}, seed: {}",
-            precommit.info.seal_randomness_height, precommit.pre_commit_block_number, interactive_block_number, hex::encode(entropy), hex::encode(precommit.seal_randomness), hex::encode(interactive_randomness));
-        log::debug!(target: LOG_TARGET, "Prover Id: {}, Sector Number: {}", hex::encode(prover_id), precommit.info.sector_number);
-
-        // Verify the porep proof
-        T::ProofVerification::verify_porep(
-            prover_id,
-            precommit.info.seal_proof,
-            commr.raw(),
-            commd.raw(),
-            precommit.info.sector_number,
-            precommit.seal_randomness,
-            interactive_randomness,
-            proofs,
-        )
-    }
-
-    /// Get randomness from the chain and process it with domain separation.
-    fn get_randomness<T: Config>(
-        personalization: DomainSeparationTag,
-        block_number: BlockNumberFor<T>,
-        entropy: &[u8],
-    ) -> Result<[u8; 32], DispatchError> {
-        // Get randomness from chain
-        let Some(randomness) = T::AuthorVrfHistory::author_vrf_history(block_number) else {
-            return Err(Error::<T>::MissingAuthorVRF.into());
-        };
-
-        // Converting block_height to the type accepted by draw_randomness
-        let block_number = block_number
-            .try_into()
-            .map_err(|_| {
-                log::error!(target: LOG_TARGET, "get_randomness: failed to convert block_height to u64");
-                Error::<T>::ConversionError
-            })?;
-
-        // HACK: convert an unsized slice to a sized one
-        // SAFETY: we know that the output is 32 bytes since we control the chain config
-        let mut sized_randomness = [0; 32];
-        sized_randomness.copy_from_slice(randomness.as_ref());
-
-        // Randomness with the bias
-        let randomness = draw_randomness(&sized_randomness, personalization, block_number, entropy);
-
-        Ok(randomness)
     }
 }
